requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pycryptodome==3.19.0
redis==5.0.1
pydantic==2.5.2
